More air dates chaos.
After seeing downsides of dynamically calculating original air date
as min over all air dates, or hardcoding that in the database (but
complicated cases when creating and deleting air date instances),
decide to just keep the notion of original air date separate from
rerun air dates. This is the beginning of the roll back.