Upgrade to sorl-thumbnail==12.4.1
Note: thumbnail.0001_initial migration was not in my django_migrataion
table, so after this upgrade I was prompted to apply the migration, which
then errored out because the thumbnail_kvstore table already exists.

Probably related to the issue here:
https://github.com/jazzband/sorl-thumbnail/issues/458

To fix, I just manually added a row to django_migrations to indicate
that this migration was already applied.